feat(derive): trim macros
